# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Url shortener app using typescript, mongoose, node.js",
    "title": "Url Shorten",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/api/shorten/": {
      "post": {
        "description": "",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "longUrl": {
                    "type": "string",
                    "required": true
                  },
                  "alias": {
                    "type": "string",
                    "required": false
                  },
                  "topic": {
                    "type": "string",
                    "enum": ["acquisition", "activation", "retention"],
                    "default": "acquisition",
                    "required": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/shorten/{alias}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/analytics/topic/{topic}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "topic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/analytics/overall": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/analytics/{alias}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "error": {
                  "example": "Not Found",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "description": "Resource not found."
      },
      "ServerError": {
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "error": {
                  "example": "Internal Server Error",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "description": "Internal server error."
      },
      "ValidationError": {
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "error": {
                  "example": "Validation Error",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "description": "Validation error."
      }
    },
    "securitySchemes": {
      "JWT": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}